requirements.txt
python-telegram-bot==20.6
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
